# Optimize groupby